some delete endpoint optimizations (#115)

+++ api/routes.py
@@ -284,13 +284,12 @@ def delete_photos(current_user: Token):
                 service.delete_user_photos_and_metadata(current_user, to_delete_user_id=user_id)
 
                 return "", 200
-            else:
+            elif current_app.config["IMG_STORAGE_PATH"]:
                 if user_id != "":
                     service.delete_temporary_user_data(user_id)
                 else:
                     service.delete_temporary_user_data(current_user.user_id)
-
-                return service.proxy_delete(current_user, user_id)
+                return service.proxy_delete_if_not_exists(current_app.config['SIMILARITY_SERVER'], current_user, user_id)
         except exceptions.MetadataNotFound as e:
             _log_error(current_user, e)
